Updates login and logout functions.
Both login and logout functions on enter and exit were changed so that Voicemeeter remote can be called from functions and outside of the with statement. This would be needed since once logout is sent, you can not send another login from the same application, you would have to restart said application.