Update util.py
When using a default argument in a python function that is a mutable type (e.x. a list, or a dictionary), python will make one version, and then when the function is called again, it will re-use the same object, causing possible unintended side effects. currently, the function is used in a way that will not cause an error, but if it is used in the future, then it can't be abused in such a way